+++ schemas/product.py
@@ -4,11 +4,11 @@ from pydantic import BaseModel
 class ProductRequest(BaseModel):
     name: str
     category: str
-    price: float
+    price: int
 
 
 class ProductResponse(BaseModel):
     id: int
     name: str
     category: str
-    price: float
+    price: int
